Add hash_map::try_emplace() / tweak hash_map::insert_or_assign()
And use it to optimize CliComm message passing.

The 'CliComm' mechanism allows external applications to subscribe to
'openMSX changes'. For example when a setting value changes, or when a
disk is inserted.

To avoid sending redundant message to other processes, we first check
that, for a specific type and topic, the value has actually changed.
This is implemented via a hash_map. For example when a setting was
"changed" we lookup the setting-name in the map and compare the
associated value (which is the last send message) with the current
setting-value. Only if they are different we actually send the message
(and update the map).

Maybe "changing" a setting into the same value is not very common, but
e.g. changing a LED status to the same value is very common (e.g. the
PPI port C gets written to, but bit 6 (the CAPS LED) keeps the same
value).

So at some point we need to test if the key is present in the map, and
if not then insert that key(+value) now. This patch makes a small tweak
to how this is done. Before we hashed the key twice: once to see if it's
present, and (if not) again to insert it. After this patch it's only
done once.

Hashing a string is not terribly expensive but it's better anyway if we
can avoid doing the same thing twice. And this tweak is located purely
inside the hash_map implementation. So it's not that we have to
complicate the user code for this.