Tweak OSDWidget::getProperties()
OSD widgets (rectangles, text) have various properties like position,
font (for text widgets), image filename (for rectangles), ...

It's possible to query the list of possible properties for a widget.
(This is mostly used interactively for tab-completion, but could also be
queried by a script via the 'osd info' command).

This list only depends on the type of the widget, not on the actual
instance. In other words: two instances of the same type will always
return the same list.

Before this patch the OSDWidget::getProperties() method would allocate a
new vector for the result. But because this result is statically known
(per type) we can tune the implementation by already calculating the
result at-compile-time and then only return a reference (implemented via
a 'span') to this pre-calculated result. This saves a vector
allocation/copy/deallocation.