Tweak MSXDevice::getName()
This patch saves ~150 string copies on openMSX startup (for a 'regular'
MSX 2 machine, with a Boosted machines it saves more like 500 copies).

MSXDevice has a method
    virtual std::string getName() const;
During startup this method is called quite often. For example:
* To construct names for sub-entities (e.g. embedded ROM, RAM,
  sound-device, debuggables) within the device.
* To ensure that the device names are unique (e.g. when inserting the
  same extension twice) the name is checked against all other names.
  (Quadratic algorithm, but OK since there aren't too many devices).

Even though MSXDevice has an internal
    std::string deviceName
member variable, the getName() is virtual and returns a string by-value
instead of returning a reference to the 'deviceName' member variable.

The only reason for this is the MSXWatchIODevice and MSXMultiDevice (and
subclasses). For example internally openMSX assumes there's exactly one
MSXDevice per (input or output) IO port. When there is no device
registered for a certain port we instead register the DummyDevice (which
returns 0xff on read and does nothing on write). And when we want more
than one device to react to a certain IO port, we instead register one
MSXMultiIODevice that then delegate to the list of devices. The name of
such a multi-device is a combination of the names of the devices in the
list. So in that case we cannot return a reference to an existing
string.

Though the above situation (having a calculated name) is pretty rare.
This patch makes a trade-off:
- Have (slightly) more complex code for the calculated-name case.
- Avoid string copies for the common case.

More in detail. This patch changes the getName() method to:
    virtual const std::string& getName() const;
So it's still a virtual method, but it now returns a reference (thus
avoiding the copy). The default implementation simply returns a
reference to the internal 'deviceName' member variable. But some classes
override this method to first update that 'deviceName' variable.

This patch avoid many string copies, and thus results in a tiny bit
faster startup speed. Though another motivation for this patch is to
enable the optimizations in the next patches (tweaks to tab-completion).