Use a 'view' instead of a 'vector' for tab-completion
Tab-completion in the openMSX console works like this (I'm skipping over
various details):
* The console code splits the line into tokens.
* This first token in the list is interpreted as the name of a command.
* We lookup the command with that name, and if it exists we delegate to
  that command's tab-completion routine.
* This routine can inspect the tokens and then do (context-sensitive)
  completion on the last token.
* Most routines will use the helper function completeFileName() or
  completeString(). This patch is about the latter.
* completeString() takes as 1st parameter the to-be-completed token and
  as 2nd parameter a list of possible values for that token.
* Internally completeString() then filters the possible values on the
  partially typed last token, and then either shows the list of
  remaining possibilities or (partially) completes the token.

This patch is about how this list of possibilities is calculated. For
example take the 'plug' command. When we need to complete the 2nd token
(=the connector name), before this patch, we would create a vector of
strings containing all possible connector names. This requires
allocating (and possibly re-allocating) a vector and making several
string copies.

Internally completeString() mostly only filters the given list, so
technically it doesn't really need a 'vector<string>'. A range of
string-like objects is sufficient. And actually we already made use of
this liberty: completeString() accepts a generic (templatized) 'range'
object. For example this could be a 'view::keys()' object, that is if we
have a 'map<string, ..>' we don't need to copy the keys of this map into
a vector, instead we (conceptually) pass (a reference to) the map plus a
transformation-function that extracts the key from the elements of the
map.

In this patch we continue the above idea. For example the 'plug' command
has access to a vector of pointer-to-Connector-objects. Before this
patch we would first fully transform this into a 'vector<string>'
containing all connector names. After this patch we instead pass:
    view::transform(connectors, [](auto* c) { return c->getName(); })
So conceptually we directly pass (a reference to) this vector of
pointers plus a transformation function that turns such a pointer into a
name. So we don't allocate any extra memory and we don't make any string
copies.

I took the 'plug' command as an example, but there are many similar
cases that got optimized in this patch. Next to tab-completion this
optimization is also beneficial for some other cases. For example
querying the list of connector names via the 'openmsx_info romtype'
command: We can use a view::transform() to add all names into a Tcl
list without constructing a temporary 'vector<string>'.