Fix swapped red/blue colors in laserdisc with SDLGL-PP
This is a regression introduced in 36cb30f3f553719e6. That patch
reverted from requiring openGL 3.3 back to openGL 2.1. See that commit
for lots more details. But part of that patch was a preparation to
support openGL ES. OpenGL ES only supports the 'RGBA' pixel format,
before we were using 'BGRA' (so red and blue swapped).

In 36cb30f3f553719e6 we forgot to adapt the laserdisc rendering for this
color channel swap. More specifically the YUV->RGB color conversion
routines (in yuv2rgb.cc). This routine is used for both the SDLGL-PP
(openGL) renderer and the SDL renderer. To avoid having to duplicate
this routine, I now also switched the SDL renderer to use the 'RGBA'
pixel format (it was still using 'BGRA').

I verified that:
* MSX screen rendering
* laserdisc rendering
* OSD rendering
* console rendering
* taking screenshots
* recording videos
Now all use the correct colors. Both in SDL and SDLGL-PP renderers.
Did I still overlook something?