Replace vector<byte> with IterableBitSet<256>
In MSXDevice we have two member variables:
    inPorts  and  outPorts
These contain the list of (input or output) port-numbers that this
device occupies.

Before we stored these numbers in an (unsorted) vector. Even an empty
vector takes 3 pointers (24 bytes). When the first element is inserted
in this vector (with the default allocator, on linux with libstdc++,
...) we allocate 16 more bytes. This patch changes the storage to an
IterableBitSet<256> which always takes 32 bytes.

So for a device with no IO ports the new representation may take 16
bytes more. For a device with only In or only Out ports the memory usage
is exactly the same. For a device with both In and Out ports the new
representation takes less memory (all assuming no device occupies more
than 16 ports). All this to say that memory usage before/after is not an
issue, basically not changed.

What is changed is memory locality and the number of heap allocations.
When starting some of the Boosted MSX machines, we now perform +70 less
heap allocations. (This is more than the number of MSXDevices with IO
ports, but that's because the vectors may need to reallocated because of
repeated push_back() operations).

Before this patch there was a problem when (in the config file) a single
device registers the same IO port more than once. Actually I think it
would work fine, just less efficient (a needless extra indirection each
time those IO ports were accessed). As a nice side-effect this patch also
fixes that problem (IterableBitSet automatically filters doubles).