Tweak multi-line OSD text rendering
More specific: use StringOp::split_view() instead of StringOp::split()
to split the to-be-rendered text on newlines.

This makes a trade-off:
 * This doesn't need any heap-allocation to store the result (because
   it's calculated on-demand).
 * We do two passes over this result. So with a view this means we
   effectively do the work of locating the newline characters twice.
   (But in practice we won't render lines with thousands of characters,
   so this still seems cheaper than heap-allocation).

This removes the last use of StringOp::split(). And this was the last
function in StringOp which performed heap-allocation.

Note: I believe multi-line text rendering (string splitting, wrapping,
joining) can be optimized further. But let's not do too much at once in
this patch.