Reduce memory usage in the 'CheckedRam' class
For the common case that the UMR checker is not active.

'CheckedRam' contains a 'Ram' object for the actual storage of an
emulated RAM chip. Next to this it stores one extra bit for every byte
in the RAM. This bit is used to track if this byte has ever been written
to. This is used by the 'Uninitialized Memory Read (UMR)' debugging
tool.

On top of this array of shadow bits, we keep one bit per 'cache line' (a
cache line in openMSX is an aligned range of 256 bytes). This extra bit
indicates whether all bytes in this cache line have been written to.
This extra info allows to more quickly decide if the RAM-access-bypass
mechanism of our CPU emulation can be used: if all bytes in a cache line
have already been initialized, then we don't need to further update the
tracking bits (on write) or check to print a warning (on read).

Typically the UMR-checker is not active. In that case we initialize the
2 helper arrays as-if all bytes have been written. And then 'CheckedRam'
has minimal overhead compared to plain 'Ram'.

If the helper array with cache-line-granularity is set to fully
initialized, then the array with byte-granularity won't be accessed.
Before this patch we would still allocate and initialize that array.
This patch avoids that.

As said at the start: that array keep one bit per byte. So for example
with a 4MB memory mapper this patch saves 512kB.