Fix(?) #1212 V9990 BMXL and BMLX command P1
With this fix the test cases attached to
  https://github.com/openMSX/openMSX/issues/1212
works fine. (Though I only needed to fix the BMXL command).

The problem description in the ticket put me on the wrong track: it
talks about BMXL/BMLX wrongly being byte-based instead of pixel-based.
But these commands are a mix of byte- and pixel-based, and it was
already implemented like that in openMSX.

The thing I fixed was that in 4bpp modes, the high and low nibbles of
the input byte were applied on the wrong pixel when the destination
rectangle starts at an odd x-coordinate. (And something similar for 2bpp
modes). More in detail: the source was correctly read as bytes, the
destination was correctly written as pixels, but there was a missing
'shift' to move the nibble from the input to the correct nibble-position
in the output.