Better use 'message' instead of osd::display_message
'message' is implemented in C++ and also outputs to other places. These places
are used for several purposes, e.g. logging the error messages for later
rereading. There is a Tcl callback for stuff that is communicated with
'message' and the default implementation maps to osd::display_message. So,
instead of using that implementation directly, let's use the more native
'message' call, so that we also get the advantages of that.

Do note that there's one thing not really working well here and that's a hack
in _osd_menu.tcl to display ROM info in a nicely formatted way. That hack is
using the OSD widget of osd::display_message directly to append another widget
with the actual content of the ROM info. That hack still works as before, but
the data doesn't end up in the 'message' stream. Still, that doesn't mean this
commit is an improvement, so that's why I put it through.

This change was triggered by the discussion here: https://www.msx.org/forum/msx-talk/openmsx/openmsx-0150-warnings-pop-ups-dissappear-too-soon