Added source for a tsx2wav conversion tool in Contrib/tsx
Based on the work done by 'NataliaPC':
    https://github.com/nataliapc/openMSX_TSX

The most important changes are:
* Fixed several buffer-overflow problems when parsing invalid tsx files.
  And made parsing more robust in general.
* Rewrote some parts to better fit the openMSX coding style. Avoid some
  code duplication.
* During this refactoring I found other (smaller) bugs. E.g. stuff
  that wasn't implemented yet, or not exactly according to the TZX
  specification. I discussed these bugs with 'NataliaPC' and 'imulilla'.

(For now?) this is only a TSX->WAV conversion tool. But it's written in
such a way that the parser code can be reused for native TSX support in
openMSX (and I've donated such an implementation to 'NataliaPC' and
'imulilla').