Implement the differences between OPM and OPP
No one was able to test this, but the behaviour was confirmed by Grauw on real
hardware. So let's then assume this implementation is also correct. Just for
the sake of having it done... It seems no one really cares if no one can test
it. But I'm tired of keeping this patch just on my hard drive, letting the time
spent on it getting wasted.

It would be nice if still someone (@gzaffin? @kumokosi?) can test/confirm that
it works now like the real hardware, i.e. software written for SFG-05 works as
intended now.

Closes #1081
Closes #1369