Fix kanji font ROM sha1sum.
Previous sum was from a bad dump. Apparently on Peter's machine, the dumping
goes wrong. We noticed when we saw it differed from the model 2 kanji font ROM,
which was a very standard one. Then we asked Jipe to check with his model 1 and
he got exactly the same as the model 2 of Peter. After redumping Peter's model
1, the differences looked very strange: there was only a few dozen bytes
different around every 16kB offset. I guess the chip is not 100% working
anymore on that machine. At least we now know that it should just be the
commonly used ROM.