Bugfix in DirAsDSK::getChainStart()
This bug was introduced in b0b3d6bcd12 (2020/12/13).
That patch replaced loops like:
  for (unsigned i = start; i < stop; ++i) { ..
with
  for (auto i : xrange(start, stop)) { ..
But that's not correct when 'i' is changed in the body of the loop.

Discovered purely by accident while playing with a new version of
clang-tidy (but older versions would likely have found it as well). More
specifically the pass "clang-analyzer-deadcode.DeadStores" reported that
in the above loop the variable 'i' was written to but later that
overwritten value is no longer used.

I checked the other code for similar bugs, but found none. But this did
trigger me to run many more clang-tidy checks on the code. None found
any additional bugs, just some (minor) code cleanups. I'll add those in
separate patches.