Implement a frame offset for the sprite locator.
As I wanted to use the frame time, I generalized that proc from the _tas_tools
script and put it in the VDP script as a generic internal function.

Also, I changed from return value to message to pass the message... This way
the script is usable from the OSD menu in the Toys and Utilities submenu.