Added new (partial) unicodemap for Arabic.
As devised by Rebecca Bettencourt and Wessam Bahnassi.

So far it only supports international mode (boot with CTRL) properly, but it's
better than assuming the international mapping, which is quite different.
Also, a part of the character set is for Right-To-Left, and that is not
supported. Only the Left-To-Right characters are mapped for now.