Split the br_hotbit unicodemap, as the charset is different.
Only on one place though, but still. Let's be accurate :)
So, the old hotbit is the hotbit_1_2. The hotbit_1_1 has a Pts symbol on 0x9E
and the hotbit_1_2 has the standard ABNT MSX-BR Cz/Cr symbol, which is also in
the character set of the "Gradiente 1.1" systems (but these have a different
matrix)..