Implement MSX character to unicode (and vice versa) mapping.
Thanks to Wouter who wrote most of the patch.

With this mapping, we can now make a proper translation from screen data to
unicode string, as is now used in the get_screen Tcl proc.

The conversion in the other direction can be used to improve type_via_keybuf,
for instance. That can be done later.

As indicated in the readme file, all character mappings are created by Rebecca
Bettencourt and they are used directly as they came from the Unicode.org
website, except that they were converted to the line endings fitting my system.

This commit closes #1397.