CPUCore: Correct R800 block I/O instruction flags.
On the Z80, contrary to the Z80 manual, block I/O instructions affect all flags
rather than just Z and N. This is described in The Undocumented Z80 Documented,
and emulated correctly.

The R800 however follows the official Z80 documentation, and block I/O
instructions only affect Z and N. It does so in exactly the way the Z80 manual
describes: Z is set if B is 0 after the instruction, and N is always set.

Fixes #1406.