Apply Z80 CMOS behaviours to all MSX-ENGINE machines.
Since all machines with MSX-ENGINE exhibit these behaviours, rather than only
the MSX turbo R. Add device definitions to all machines with T7937, T7937 and
T9763 engines in the process.

The “out (c),0”, “scf” and “ccf” instructions behave the same on my Panasonic
FS-A1FX and Panasonic FS-A1GT behave the same, and so does the Sanyo MPC-2
according to the comment in the xml. Common denominator is that all of them have
a CMOS Toshiba MSX-ENGINE with embedded TMPZ84C00A Z80. This is related to CMOS
tech interacting differently with floating buses internally in the Z80, see:
https://sinclair.wiki.zxnet.co.uk/wiki/Z80

Fixes #1408.