Workaround visual studio bug(?), remove semiregular.hh
For some reason it triggers compile errors in visual studio in c++20
mode, but not in c++17 mode. Possibly there's a conflict with the c++20
std::semiregular concept?? Even though both live in different
namespaces. Seems like a bug in visual studio?

We could probably come up with some workaround. (Maybe rename the
type???) On the other hand, I checked, and at the moment the
'semiregular' helper isn't needed anymore(*). It's only used in our
implementation of view::transform(), and in the (near?) future we can
switch to std::views::transform(). So instead of trying to fix the
visual studio compile error in 'semiregular', we can also remove it.

(*) It's only needed for certain instantiations of view::transform.
Probably the code that once triggered those instantiations got
rewritten.