Fix crash on "record_channels cassetteplayer" with resampler=blip
Some background info:

Sound devices in openMSX, in principle, generate the output of their
different channels in separate buffers. This allows stuff like
per-channel recording, mute, balance, ... Later those channels are mixed
(=summed) together.

Though very often there is no per-channel recording active, there are no
channels muted, all channels have the same balance, ... In that case
generating the output in different buffers requires more memory, and
mixing requires an extra pass over all the data. Both are bad for
cache-locality.

To optimize this common situation, the sound devices must, instead of
simply generating each channel in a buffer, add the data of a channel
into a buffer. In the common case of "no separate output needed",
there's only one buffer that is pre-filled with zeros, and the output of
each channel is added into this buffer. In the less common case that
(some) channels are needed separately, there are separate (also
pre-filled with zeros) buffers for those channels.

Adding data into a buffer is only slightly more expensive than simply
overwriting the data in the buffer (it's 'load-add-store' compared to
only 'store'). So this is a win for the common case.

For single-channel sound devices the above optimization doesn't make
sense. For these we just directly store the (single) channel data into
an (uninitialized) buffer.

The bug:

* For single-channel sound-devices we don't zero-initialize the output
  buffer, because that data is anyway overwritten (instead of added to).
* In case separate channels are needed, those channels are first
  generated in separate buffers and later mixed with the other channels.
But this mixing code didn't take the uninitialized-single-channel-buffer
stuff into account. It mixes (=sums) into an uninitialized buffer, and
that results in garbage.

This explains why recording a single-channel sound device results in
garbage. It doesn't yet explain why it crashes when using the blip
resampler.

The blip-resampler is optimized for the case that many consecutive
output samples have the same value, so output that resembles square
waves. This is often the case with PSG and SCC sounds.

In the inner loop of the blip-resampler we have to test two conditions:
* Has the new sample the same value as the previous one.
* Did we reach the end of the buffer.
By placing a sentinel at the end of the buffer we can fold both into a
single check. That is: only when two consecutive samples differ we check
that we have reached the end of the buffer. This makes the common case
of identical consecutive samples twice as fast.

The sentinel is given the value "last-sample + 1", so we should always
detect a difference at the end of the buffer. However, samples are
stored as floating point values (initially was used integers, but since
a few years we've switched to float). Because of the above bug it's
possible those samples are uninitialized values, so by accident it could
be a floating point value like "1e40". Adding '1' to such a large value
doesn't change that value, and then the sentinel trick doesn't work
anymore. We then read/write past the end of the buffer and eventually
crash.

The fix:

The code-path that requires separate channels (e.g. because recording is
enabled) is anyway a slow-path through the sound-code. In that path we
now always zero-initialize the output buffer, thus also for single
channel sound devices. In theory we can avoid this, but it's anyway the
slow-code-path. It's good enough.