Use c++20 concepts to constrain 'template <typename T>'
A template function like:
    template<typename T>
    void f(const T& t) { ...
can in principle be instantiated with *any* type. Though most often the
function won't actually work with really *all* types. Maybe it only
works with integral types (as in: any integral type, not just 'int').
Or maybe T is supposed to some kind of 'callable' object.

So often there are constraints on types (still plural) that can be
accepted by the function. In other words: 'T' is supposed to model some
'concept'. See here for lots more details:
    https://en.cppreference.com/w/cpp/language/constraints

For example the above function can be constrained like this:
    template<std::unsigned_integral T>
    void f(const T& t) { ...
Now the function can only be instantiated with unsigned integral types
(like 'unsigned char', 'unsigned short', 'unsigned int', etc).

Concepts can also be combined with the abbreviated template syntax (see
previous patch) like in this example:
    void f(std::predicate<int> auto p) { ...
Now the function can only be instantiated with types that satisfy the
'predicate<int>' concept (that is: it's invocable with a single 'int'
parameter, and it returns a bool result).

One of the main benefits of concepts is that it explicitly states
constraints on the allowed types (possibly before these were stated in
comments, but the compiler doesn't read comments). (Often) this allows
the compiler to generate clearer error messages when some of the
constraints aren't satisfied. But see later patches for other benefits.