Use c++20 concepts to simplify absHelper/putSignHelper
The previous patches used c++20 concepts to constrain the allowed types
in templates (and then the compiler can generate more descriptive error
messages on programming errors).

This patch uses c++20 concepts to simplify function overloading. More
specifically, depending on whether a type is signed or unsigned, we want
to call one or another implementation of a function. Before we
implemented that via a struct templatized on a bool with two different
specializations. Now we simply implement two function and constrain one
to signed and the other to unsigned types. See the patch for details,
that's probably clearer than my attempt to explain it here.

And in some cases using 'if constexpr' is even simpler (that was already
possible in c++17).