Use c++20 concepts to constrain 'template<typename Pixel>'
Many of the rendering classes/functions were templatized on
    template<typename Pixel>
but instead of accepting any type 'Pixel' was (implicitly) supposed to
be either uint16_t or uint32_t (respectively for 16bpp or 32bpp pixel
formats). This patch constrains the allowed types to
'std::unsigned_integral'.

Note: is it still useful to support 16bpp pixel formats? Some years ago
we still supported handheld systems like 'dingoo' those were still
16bpp. In the future we might do the software part of the MSX rendering
in 16bpp (sufficient for MSX) and then use hardware (=GPU) scalers to do
the color transformation to 32bpp (currently this transformation is done
via a LUT on the CPU).