Generate update-events on breakpoint/watchpoint/condition/probe add/remove (issue#1414)
This is meant for external applications controlling openMSX. So in this
case most likely a debugger ;-)

To use it, the application must first send the command:
    openmsx_update enable debug
After that the application may receive (asynchronous) update events
like:
    <update type="debug" machine="machine1" name="bp#1">add</update>
(Similar to how adding/removing hardware events are structured).

Consider this experimental for now. We still need to gain implementation
experience by actually using this in an external application.