Remove filter from all but LED and SETTING update-events
OpenMSX can send update-events to external applications (catapult,
debugger, ...). Examples are: some LED has turned on/off, a different
disk was inserted, ...

To avoid sending duplicate messages we filter them: e.g. don't send that
the CAPS LED has turned on, when we already send that in an earlier
message (I mean for the same MSX machine instance, for the same LED,
without an intermediate CAPS-LED-off message).

This filter was implemented in a generic way. Meaning that all update
messages got this filter. Though, by design, most update-event-types
won't repeat the same message. The only exceptions are LED events and
SETTING-change events. This patch makes the filter optional.

--

I 'discovered' this while working on the events for break- and
watchpoints. When adding/removing breakpoints, openMSX always creates
new names (bp#1, bp#2, ...). This meant that (the internal map in) the
event-filter would fill-up with more and more new breakpoint names, but
it never does something useful. (Because all names are new, there's
never a repeating message).

For other event-types (like plug-events, hardware-added/remove events),
the filter is also useless, but at least the memory usage would not keep
on growing.

Anyway, this patch renames the old update() method (with filter) to
updateFiltered() and introduces a new update() method (reusing the old
name) which bypasses the filter. Only for LED and SETTING events we
switch to the updateFiltered() version.