Fix compiler warning on gcc-12
Gcc-12 gave a warning about potential use of uninitialized variable.
Fixed by swapping the order of the NowindInterface 'host' and 'drive'
member variables.

Initially 'drives' is an empty std::vector. 'host' takes a reference to
'drives' as constructor parameter and stores this reference (it doesn't
use it yet).

Before this patch we would first construct 'host', thus pass it a
reference to the not-yet-initialized 'drives'. Only later we would
initialize 'drives' (as an empty vector). Currently that's harmless
because at that point 'host' doesn't use 'drives' yet. But it's better
to fix it anyway.