Fix debug-build after cabc54edc9
Revision cabc54edc9 added the enum value CliComm::UpdateType::DEBUG, but
in a debug build we set the macro 'DEBUG' and this causes a conflict.

For now fixed by renaming the enum value to DEBUG_UPDT.

A better fix would be to eliminate the use of the DEBUG macros (in favor
of a boolean constant), or when this is not always possible, rename the
macro to a less common name. Though this is more invasive, and I didn't
want to do this close before the release.