Use c++20 concepts to replace 'typename Archive'
For load/savestates many classes implement a method like:
    template<typename Archive>
    void serialize(Archive& ar, unsigned version) { ...
In principle this method could be instantiated with any type, though the
intention is to only do this for types that implement an 'Archive'
(according to the meaning of 'Archive' for our serialization stuff).

More concretely it should only be instantiated with one of these 4
types:
    XmlOutputArchive   (for saving to an XML file)
    XmlInputArchve     (for loading from an XML file)
    MemOutputArchive   (for saving to memory)
    MemInputArchive    (for loading from memory)

In this patch we create a custom c++20 concept called 'Archive' (the
implementation of this concepts checks that it's one of the 4 above
types). And then we can define the serialize() methods like this:
    void serialize(Archive auto& ar, unsigned version) { ...
This constraints the template to only types that make sense (and it's
also one line shorter).