Add support for 2nd PSG.
More and more game enhancement patches make use of it, so let's add a specific
extension and add it to the Boosted configs. Note that the I/O port does
confict with the Philips NMS 1205 Music Module, but only with that specific
extension. The other MSX-AUDIO (like) configs do not have these MIDI I/O ports
there.