Fix interaction between VDP command engine and CPU VRAM-access-slots (fixes #1422)
When both the VDP command engine and the CPU are simultaneously
accessing VRAM, the CPU gets priority. Thus the command executes
slightly slower. Though because of a bug the emulated command was
(sometimes) actually executing faster rather than slower.

More in detail: the bug was in VDPCmdEngine::stealAccessSlot().
This method allows the CPU to steal an access slot from the command
engine. It does this by advancing the engine-time to the first
available access slot that's past the current CPU time. Unfortunately
that's wrong. This is because "engine-time" is not the last time at which the
command engine accessed VRAM. Instead it's the next (future) time at
which the command engine wants to access VRAM. So instead of using the
CPU-time, we must use the engine-time as the start-point to advance to
the next available access-slot.

How is this related to this bug report?
    Replay desyncs when using higher speed than 100 #1422
    https://github.com/openMSX/openMSX/issues/1422
To avoid having to synchronize with the VDP command engine on every
status register read (because typically the CPU reads this register in a
tight polling loop) we make a (lower bound) estimate for when the VDP
command will finish (because before the command finishes, the status
register won't change). Though because of the above bug (=command
executes faster than expected) our lower-bound estimate was actually
higher than the actual (but wrong) time at which the command finishes.
Because of this we did too few sync, and the value of the VDP status
register value was not 100% reliable: it now also depends on extra syncs
e.g. triggered by frame rendering. And running at different speeds may
cause different frameskip behavior.