c++20 can synthesize operator!= from operator==
Practically always operator!= is implemented like this:
    bool operator!=(const A& x, const A& y) { return !(x == y); }
c++20 knows that operator!= is supposed to be the inverse of operator==.
When operator!= is not explicitly implemented, c++20 will synthesize it
from the corresponding operator==. This reduces the amount of
boilerplate code.

There something similar for heterogeneous comparisons:
    bool operator==(const A& x, const B& y) { ...
c++20 understands that operator== is supposed to be commutative. So from
the above function, the compiler will synthesize a version with the
order of the 'A' and 'B' parameters swapped. And combined with the
above, it will also synthesize 'A!=B' and 'B!=A'.