c++20: request default operator== when all members need to be compared
Often we have a simple struct with a operator== that compares all the
members. For example:
    struct S {
        std::string s;
        int i;
        float f;

        bool operator==(const S& other) const {
            return (s == other.s) &&
                   (i == other.i) &&
                   (f == other.f);
        }
    };

c++20 can synthesize such operator== functions upon request (similar to
how it already can synthesize a copy-constructor and copy-assignment for
such simple structs). So the above function can be replaced with:
        bool operator==(const S&) const = default;

In a similar manner, c++20 can synthesize operator<, <=, >=, > for
simple structs: this does a lexicographical compare on all members.
More in detail: c++20 can synthesize operator<=> (the spaceship
operator) and from this <, <=, >=, > can be obtained. More on
operator<=> in the next patch.