Replace M_PI with Math::Pi
And similar for M_E, M_LN2, M_LN10.

Athough M_PI is more compact, it's not fully portable. E.g. it's
difficult to use with visual studio (depends on order of includes).
For this reason we had to define these constants ourselves.

c++20 added the <numbers> header
    https://en.cppreference.com/w/cpp/header/numbers
and this (finally) gives a portable interface to various mathematical
constants. We still wrapped it in our 'Math.hh' header to make it a bit
more convenient to use.