Fix assertion when running with cmdtiming=broken
Follow-up on f898607b3ef. That commit would advance the cpu-access-time
to 16 VDP cycles in the future in the VDPCmdEngine::stealAccessSlot()
method. Though the VDP itself already took care of this (via the
VDP::scheduleCpuVramAccess() stuff).

A side effect of this fix is that the VDP commands will execute slightly
faster again during simultaneous CPU VRAM access (was a regression
introduced in f898607b3ef).