Fix desync in replays with mouse input
To translate from host- to msx-mouse-input we scale down the x/y-offset
by a factor 2. Though we keep track of rounding errors. So that e.g. two
consecutive host offset of '7 and 7' result in '3 and 4' msx-offsets.
Unfortunately we forgot to include this rounding status (when to round
up or down) in the savestate. This can give (slightly) different MSX
mouse input and in the end cause a desync.

Instead of fixing this by including this 'rounding-status' in the
savestate, we fix this by not storing host but msx mouse-offsets in the
replay file. Thus doing to host->msx translation sooner. Thanks mth for
pointing this out.