tweak host->msx mouse-offset translation
To translate host->msx mouse movement, we divide the host offset by two
(and keep track of rounding errors, see previous patch). Though in C/C++
division is required to round towards zero. So a different direction for
positive and for negative values. For a smoother output it's better to
use the same rounding direction for all values.

In practice it's hard to tell the difference between the new and the old
implementation, but it should be a little better now :)